% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_species.R
\name{add_species}
\alias{add_species}
\title{Add Species to Tree and Update Package Data}
\usage{
add_species(
  species,
  input_nwk_path = "trees/newick/speciesTree.nwk",
  output_nwk_path = "trees/newick/speciesTree_proof.nwk",
  output_rda_path = "data/speciesTree_proof.Rda"
)
}
\arguments{
\item{species}{A character vector of species names to add (e.g., c("Bazzania magnifica", "Mnioloma sp. 1"))}

\item{input_nwk_path}{Path to the input Newick file. Defaults to "newickTree/speciesTree.nwk"}

\item{output_nwk_path}{Path to the output Newick file. Defaults to "newickTree/speciesTree_proof.nwk"}

\item{output_rda_path}{Path to the utput Rda file. Defaults to "data/speciesTree_proof.Rda"}
}
\value{
Inserts the tips, write the files, and returns the updated phylo object invisibly.
}
\description{
Add Species to Tree and Update Package Data
}
