% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_by_genus.R
\name{tree_by_genus}
\alias{tree_by_genus}
\title{Prunes a phylogenetic tree to include only a specified set of genera}
\usage{
tree_by_genus(genus, input_tree)
}
\arguments{
\item{input_tree}{An object of class 'phylo' representing the input phylogenetic tree.}

\item{genus_to_keep}{A character vector containing the names of the genus
that should be retained in the new tree. These names must exactly match
tip labels in the input_tree.}
}
\value{
A new 'phylo' object containing only the specified genus, or NULL
        if an error occurs (e.g., no common genus found).
}
\description{
Prunes a phylogenetic tree to include only a specified set of genera
}
