% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_by_species.R
\name{tree_by_species}
\alias{tree_by_species}
\title{Prunes a phylogenetic tree to include only a specified set of species}
\usage{
tree_by_species(species_vector, input_tree, tolerance = 2)
}
\arguments{
\item{species_vector}{A character vector containing the names of the species
that should be retained in the new tree. These names must exactly match
tip labels in the input_tree.}

\item{input_tree}{An object of class 'phylo' representing the input phylogenetic tree.}

\item{tolerance}{A number representing the tolerance of character edits for each element of species_vector input. tolerance = 0 means there is a perfect word match. Default is 2 (i.e. 2 character edits is the maximum tolerance).}
}
\value{
A new 'phylo' object containing only the specified species, or NULL
        if an error occurs (e.g., no common species found).
}
\description{
Prunes a phylogenetic tree to include only a specified set of species
}
\examples{
species_list <- c("Herbertus sendtneri", "Micropterygium carinatum", "Lepidozia pinnaticruris", "Bazzania pallidevirens", "Bazzania jamaicensis", "Plagiochila simplex", "Plagiochila revolvens")
output_tree <- tree_by_species(species_list)
plot(output_tree) #Graph
}
