% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_by_family.R
\name{tree_by_family}
\alias{tree_by_family}
\title{Prunes a phylogenetic tree to include only a specified set of families}
\usage{
tree_by_family(family, input_tree)
}
\arguments{
\item{input_tree}{An object of class 'phylo' representing the input phylogenetic tree.}

\item{family_to_keep}{A character vector containing the names of the family
that should be retained in the new tree. These names must exactly match
tip labels in the input_tree.}
}
\value{
A new 'phylo' object containing only the specified family, or NULL
        if an error occurs (e.g., no common family found).
}
\description{
Prunes a phylogenetic tree to include only a specified set of families
}
