% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_by_genus.R
\name{tree_by_genus}
\alias{tree_by_genus}
\title{Prunes a phylogenetic tree to include only a specified set of genera}
\usage{
tree_by_genus(genus_vector, input_tree)
}
\arguments{
\item{genus_vector}{A character vector containing the names of the genera
that should be retained in the new tree. These names must exactly match
tip labels in the input_tree.}

\item{input_tree}{An object of class 'phylo' representing the input phylogenetic tree.}
}
\value{
A new 'phylo' object containing only the specified genus, or NULL
        if an error occurs (e.g., no common genus found).
}
\description{
Prunes a phylogenetic tree to include only a specified set of genera
}
\examples{
genus_list <- c("Bazzania", "Bazzania", "Calypogeia", "Trichocolea", "Leiomitra", "Mnioloma", "Paracromastigum", "Pseudocephalozia", "Telaranea", "Zoopsidella")
output_tree <- tree_by_genus(genus_list)
plot(output_tree) #Graph
}
