% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_by_species.R
\name{tree_by_species}
\alias{tree_by_species}
\title{Prunes a phylogenetic tree to include only a specified set of species}
\usage{
tree_by_species(species, input_tree)
}
\arguments{
\item{input_tree}{An object of class 'phylo' representing the input phylogenetic tree.}

\item{species_to_keep}{A character vector containing the names of the species
that should be retained in the new tree. These names must exactly match
tip labels in the input_tree.}
}
\value{
A new 'phylo' object containing only the specified species, or NULL
        if an error occurs (e.g., no common species found).
}
\description{
Prunes a phylogenetic tree to include only a specified set of species
}
